% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokedex.R
\docType{data}
\name{pokedex}
\alias{pokedex}
\title{Informations of pokemon stats up to 8th gen.}
\format{
A data frame with 1045 rows and 50 variables:
\describe{
\item{pokedex_number}{number of pokemon on the National Pokedex.}
\item{name}{name of the pokemon in english.}
\item{german_name}{name of the pokemon in german.}
\item{japanese_name}{name of the pokemon in japanese.}
\item{generation}{pokemon's origin generation, ranges from 1 to 8.}
\item{status}{indicates if pokemon is Normal, Legendary, Sub-Legendary or Mythical.}
\item{species}{pokemon identifier based on its defining biological characteristics.}
\item{type_number}{describes if a pokemon is a mono or dual type.}
\item{type_1}{first type of the pokemon.}
\item{type_2}{second type of the pokemon, if any.}
\item{height_m}{mean height of the pokemon, in meters.}
\item{weight_kg}{mean weight of the pokemon, in kilograms.}
\item{abilities_number}{describes if a pokemon has 1, 2 or 3 possible abilities.}
\item{ability_1}{first ability of the pokemon.}
\item{ability_2}{second ability of the pokemon, if any.}
\item{ability_hidden}{hidden ability of the pokemon.}
\item{total_points}{sum of all of the pokemon's stat poits.}
\item{hp}{pokemon's base hp stat value.}
\item{attack}{pokemon's base attack stat value.}
\item{defense}{pokemon's base defense stat value.}
\item{sp_attack}{pokemon's base special attack stat value.}
\item{sp_defense}{pokemon's base special defense stat value.}
\item{speed}{pokemon's base speed stat value.}
\item{catch_rate}{pokemon base catch rate used to calculate catching probabilities.}
\item{base_friendship}{friendship value when the pokemon is caught.}
\item{base_experience}{base value used to calculate how much exp a pokemon yields when defeated.}
\item{growth_rate}{describe how fast a pokemon levels up and gains experience.}
\item{egg_type_number}{how many egg types the pokemon has.}
\item{egg_type_1}{pokemon's first egg type.}
\item{egg_type_2}{pokemon's second egg type, if any.}
\item{percentage_male}{probability of finding a male pokemon in the wild of this species.}
\item{egg_cycles}{internal value used to track how long until an egg hatches.}
\item{against_normal}{value describing pokemon's received damage multiplicator against a normal type attack.}
\item{against_fire}{value describing pokemon's received damage multiplicator against a fire type attack.}
\item{against_water}{value describing pokemon's received damage multiplicator against a water type attack.}
\item{against_electric}{value describing pokemon's received damage multiplicator against a electric type attack.}
\item{against_grass}{value describing pokemon's received damage multiplicator against a grass type attack.}
\item{against_ice}{value describing pokemon's received damage multiplicator against a ice type attack.}
\item{against_fight}{value describing pokemon's received damage multiplicator against a fighting type attack.}
\item{against_poison}{value describing pokemon's received damage multiplicator against poison type attack.}
\item{against_ground}{value describing pokemon's received damage multiplicator against a ground type attack.}
\item{against_flying}{value describing pokemon's received damage multiplicator against a flying type attack.}
\item{against_psychic}{value describing pokemon's received damage multiplicator against a psychic type attack.}
\item{against_bug}{value describing pokemon's received damage multiplicator against a bug type attack.}
\item{against_rock}{value describing pokemon's received damage multiplicator against a rock type attack.}
\item{against_ghost}{value describing pokemon's received damage multiplicator against a ghost type attack.}
\item{against_dragon}{value describing pokemon's received damage multiplicator against a dragon type attack.}
\item{against_dark}{value describing pokemon's received damage multiplicator against a dark type attack.}
\item{against_steel}{value describing pokemon's received damage multiplicator against a steel type attack.}
\item{against_fairy}{value describing pokemon's received damage multiplicator against a fairy type attack.}
}
}
\source{
\url{https://www.kaggle.com/mariotormo/complete-pokemon-dataset-updated-090420}
}
\usage{
pokedex
}
\description{
A dataset containing many information regarding pokemon up to gen 8th games. First published by Mario Tormo in Kaggle.
}
\keyword{datasets}
